Initial commit to get things going
This is just a starting point for the command line interface. It uses
the standard argparse library, and this seems to do what I need so far.
It may not be the ideal choice for the kind of interface I want to
provide (something like the aws-cli project), but I'll work within its
constraints, since I don't want to have to do any of my own parsing in
this respect.