Use config provider class to get ripper path
This class can provide various different configuration information. It
will be tested in isolation, making the whole system much easier to
work with from a testing point of view (instances of the configuration
provider can be injected into classes and its calls mocked once, as
opposed to lost of OS related calls).