Refactor command base to pass config provider
This also involves refactoring the command parser to receive the same
dependencies, so that it can pass them to commands it instantiates.
This may become slightly messy, because it means the command parser
would have to have *all* the dependencies required by any commands it
instantiates. It hopefully shouldn't be too hard to deal with this by
having some kind of wireup method that configures the command parser.