Refactor to manually generate temp file
It turned out I hadn't looked closely enough at the temporary file
function provided by the standard library. It seems it's basically
intended to be used in a context manager, and won't exist outside of
that context manager. In this case, the temp file needs to exist until
the ripper has finished running.