Case for track number and track total
This will cause a whole bunch of other test cases to break.
Unfortunately, I'm unable to run the whole suite at the moment, because
the discogs API seems to start blocking requests after 20 of them. I
think I'll need to write a little bit of code that maintains a counter
and tells it to sleep after there have been 20 requests.