Code for parsing the track data for a release
When I say track data, what I mean is the track and disc numbers and
totals. It turned out that there's a lot of messy code to get this to
work properly. It involved looping over the track list more than once.
I'm not proud of the code. There's probably a better way to do this.
However, it does work, and all the tests pass.