Case for getting release by ID
This test case is incorrect. It's not just enough to check the type
being returned, because the method doesn't need to reference the ID
passed in. It can just pass any valid ID and ignore the ID passed in. I
could change the test to check the artist of the release, but I'd
prefer to add an ID to the release model.