Remove duplicate artist number
This commit actually contains a few things. There is a test case for
removing the number on both a release and on a track, but also, the
test case for a release that has artists on tracks had to be changed to
a release where there weren't any duplicate artists.