Provide target encoding
Based on the target encoding, we differentiate where the encoded files
are going to go. The config provider has been updated to look for a
setting with the target encoding prefixed, such as
mp3_releases_base_directory. This way completely different paths can be
used for different formats, if you want to store those in different
folders (which I personally do).